Round speed, accuracy and altitude

+++ src/ProjectIvy.Model/Binding/Tracking/TrackingBinding.cs
@@ -4,9 +4,9 @@ namespace ProjectIvy.Model.Binding.Tracking
 {
     public class TrackingBinding
     {
-        public double? Accuracy { get; set; }
+        public decimal? Accuracy { get; set; }
 
-        public double? Altitude { get; set; }
+        public decimal? Altitude { get; set; }
 
         public decimal Latitude { get; set; }
 
@@ -14,6 +14,6 @@ namespace ProjectIvy.Model.Binding.Tracking
 
         public DateTime Timestamp { get; set; }
 
-        public double? Speed { get; set; }
+        public decimal? Speed { get; set; }
     }
 }
